+++ pkg/goanalysis/packages/util.go
